TUI: move the settle-window test to its own module
The tracker unit test needs no socket harness, and keeping it in
test_transport_subscription.py collided with #587's edits to the same
import block; the branches now merge cleanly in either order.
