+++ tests/server/test_transport_subscription.py
@@ -3,7 +3,6 @@
 import json
 import socket
 import threading
-import time
 import uuid
 from collections.abc import Callable, Generator
 from contextlib import contextmanager
@@ -27,7 +26,6 @@ from server.events import (
     RunEvent,
     RunStartedData,
 )
-from server.transport.subscriptions import SubscriptionTracker
 from server.transport.unix_jsonl import UnixJsonlServer
 
 _TIMESTAMP = datetime(2026, 1, 1, tzinfo=UTC)
@@ -330,30 +328,6 @@ def test_disconnect_wait_blocks_until_last_subscriber_closes(tmp_path: Path) -> 
         assert unblocked.wait(timeout=5)
         waiter.join(timeout=5)
         assert not waiter.is_alive()
-
-
-def test_disconnect_wait_bridges_a_redial_inside_the_settle_window() -> None:
-    tracker = SubscriptionTracker()
-    unblocked = threading.Event()
-    first = tracker.track()
-    first.__enter__()
-
-    def wait_for_none() -> None:
-        tracker.wait_for_none_active(settle_seconds=0.4)
-        unblocked.set()
-
-    waiter = threading.Thread(target=wait_for_none, daemon=True)
-    waiter.start()
-    time.sleep(0.05)
-    first.__exit__(None, None, None)
-    # The redial lands inside the settle window, so the wait must keep
-    # blocking well past where the window alone would have expired.
-    with tracker.track():
-        assert not unblocked.wait(timeout=0.6)
-
-    assert unblocked.wait(timeout=2)
-    waiter.join(timeout=2)
-    assert not waiter.is_alive()
 
 
 def test_wait_for_change_does_not_parse_events(tmp_path):  # noqa: ANN001, ANN201
